+++ crates/bencode/src/enums/bencode.rs
@@ -2,7 +2,7 @@ use std::collections::BTreeMap;
 
 #[derive(Debug, PartialEq, Eq)]
 pub enum BencodeValue {
-    Int(i64),
+    Int(isize),
     Str(Vec<u8>),
     List(Vec<BencodeValue>),
     Dict(BTreeMap<String, BencodeValue>),
